Folding RemoteEncoder into the RemoteMarshallingOps
